Move helpers to util
